Update class name CFish to Fish
